pyproject.toml
[project]
name = "real-time-translator"
version = "0.1.0"
description = "Real-time speech translation system with Whisper ASR, MiniMax translation, and T2V speech synthesis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "openai-whisper>=20231117",
    "webrtcvad>=2.0.10",
    "numpy>=1.24.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "setuptools>=68.0.0",
    "aiohttp>=3.8.0",
]
